refactor: standardize toast imports and inline service actions menu
