+++ apps/web/src/routes/_authenticated/$slug/$teamSlug/_dashboard/services/-services/service-benefits-tab.tsx
@@ -20,7 +20,7 @@ import {
 } from "@modules/billing/services/benefits-aggregates";
 import { PauseCircle, Plus, Sparkles } from "lucide-react";
 import { Suspense, useCallback, useMemo, useState } from "react";
-import { toast } from "sonner";
+import { toast } from "@packages/ui/components/sonner";
 import { DataTableContent } from "@/components/data-table/data-table-content";
 import { DataTableEmptyState } from "@/components/data-table/data-table-empty-state";
 import { DataTableRoot } from "@/components/data-table/data-table-root";
